pyproject.toml
[tool.poetry]
name = "dspy-agent"
version = "0.1.0"
description = "An autonomous LLM agent built on Stanford DSPy"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "dspy_agent"}]

[tool.poetry.dependencies]
python = "^3.9"
dspy-ai = "^2.3.3"
typer = "^0.9.0"
rich = "^13.6.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"

[tool.poetry.scripts]
dspy-agent = "dspy_agent.cli:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
